refactor: guard not signed in user
